+++ src/Tests/Framework/ManagedElasticsearch/Clusters/INestTestCluster.cs
@@ -1,4 +1,5 @@
-﻿using Nest;
+﻿using System;
+using Nest;
 
 namespace Tests.Framework.ManagedElasticsearch.Clusters
 {
